feat: integration of pagination in products page
