+++ HTML/productos.html
@@ -35,22 +35,23 @@
     </div>
     <!-- MENSAJE DE ALERTA -->
 
-    <!-- ENCABEZADO -->
     <main class="my-3">
       <div class="text-center my-4">
+        <!-- ENCABEZADO -->
         <div class="page__title-container">
           <h1 class="page__title text-secondary font-size-md">Productos</h1>
           <span class="page__title-span w-100 text-secondary font-size-md"
             >Productos</span
           >
         </div>
+        <!-- ENCABEZADO -->
       </div>
       <!--CONTAINER  -->
       <div class="container w-90 flex-grow-1">
-        <div class="row row-cols-1 row-cols-sm-2 row-cols-lg-3 g-5" id="items">
+        <div class="row g-5" id="items">
           <!-- PLACEHOLDERS -->
           <!-- PLACEHOLDER CARD 1 -->
-          <div class="d-flex flex-column flex-grow-1 px-3">
+          <div class="d-flex flex-column px-3 col-12 col-sm-6 col-lg-4">
             <div class="card text-white text-center bg-dark" aria-hidden="true">
               <div class="card-header">
                 <h5 class="card-title placeholder-glow">
@@ -83,7 +84,7 @@
             </div>
           </div>
           <!-- PLACEHOLDER CARD 2 -->
-          <div class="d-flex flex-column flex-grow-1 px-3">
+          <div class="d-flex flex-column px-3 col-12 col-sm-6 col-lg-4">
             <div class="card text-white text-center bg-dark" aria-hidden="true">
               <div class="card-header">
                 <h5 class="card-title placeholder-glow">
@@ -116,7 +117,7 @@
             </div>
           </div>
           <!-- PLACEHOLDER CARD 3 -->
-          <div class="d-flex flex-column flex-grow-1 px-3">
+          <div class="d-flex flex-column px-3 col-12 col-sm-6 col-lg-4">
             <div class="card text-white text-center bg-dark" aria-hidden="true">
               <div class="card-header">
                 <h5 class="card-title placeholder-glow">
@@ -149,7 +150,7 @@
             </div>
           </div>
           <!-- PLACEHOLDER CARD 4 -->
-          <div class="d-flex flex-column flex-grow-1 px-3">
+          <div class="d-flex flex-column px-3 col-12 col-sm-6 col-lg-4">
             <div class="card text-white text-center bg-dark" aria-hidden="true">
               <div class="card-header">
                 <h5 class="card-title placeholder-glow">
@@ -182,7 +183,7 @@
             </div>
           </div>
           <!-- PLACEHOLDER CARD 5 -->
-          <div class="d-flex flex-column flex-grow-1 px-3">
+          <div class="d-flex flex-column px-3 col-12 col-sm-6 col-lg-4">
             <div class="card text-white text-center bg-dark" aria-hidden="true">
               <div class="card-header">
                 <h5 class="card-title placeholder-glow">
@@ -215,7 +216,7 @@
             </div>
           </div>
           <!-- PLACEHOLDER CARD 6 -->
-          <div class="d-flex flex-column flex-grow-1 px-3">
+          <div class="d-flex flex-column px-3 col-12 col-sm-6 col-lg-4">
             <div class="card text-white text-center bg-dark" aria-hidden="true">
               <div class="card-header">
                 <h5 class="card-title placeholder-glow">
@@ -249,20 +250,19 @@
           </div>
           <!-- END PLACEHOLDER -->
         </div>
+        <nav>
+          <ul
+            class="pagination justify-content-center gap-3 my-3"
+            id="pagination"
+          ></ul>
+        </nav>
       </div>
       <!-- END CONTAINER -->
     </main>
 
-    <!-- ROPA -->
-    <!-- paginación -->
-
-    <!-- fin paginación -->
-
-    <!-- ROPA -->
-
     <!-- TEMPLATE CARD Original-->
     <template id="template-card">
-      <div class="d-flex flex-column flex-grow-1 px-3">
+      <div class="d-flex flex-column px-3 col-12 col-sm-6 col-lg-4">
         <div class="card text-white text-center bg-dark">
           <div class="card-header">
             <h5 id="name">name:</h5>
