--- PAGINACIÓN/producto_paginacion.css
@@ -1,80 +1,0 @@
-* {
-	margin: 0;
-	padding: 0;
-	box-sizing: border-box;
-	font-family: sans-serif;
-}
-
-body {
-	background: #F3F3F3;
-}
-
-header {
-	display: flex;
-	justify-content: center;
-	align-items: center;
-	height: 60px;
-	background-color: #EEE;
-	box-shadow: 0px 0px 10px rgba(0, 0, 0, 0.2);
-}
-
-main {
-	display: flex;
-	flex-direction: column;
-	align-items: center;
-}
-
-main .list {
-	width: 100%;
-	max-width: 768px;
-	background-color: #FFF;
-	border: 1px solid #CCC;
-	margin-top: 25px;
-}
-
-main .list .item {
-	padding: 15px;
-	border-bottom: 1px solid #CCC;
-}
-main .list .item:last-of-type {
-	border-bottom: none;
-}
-main .list .item:hover {
-	background: rgba(0, 0, 0, 0.05);
-}
-
-.pagenumbers {
-	display: flex;
-	flex-wrap: wrap;
-	justify-content: center;
-	align-items: center;
-}
-
-.pagenumbers button {
-	width: 50px;
-	height: 50px;
-
-	appearance: none;
-	border: none;
-	outline: none;
-	cursor: pointer;
-
-	background-color: #44AAEE;
-
-	margin: 5px;
-	transition: 0.4s;
-
-	color: #FFF;
-	font-size: 18px;
-	text-shadow: 0px 0px 4px rgba(0, 0, 0, 0.2);
-	box-shadow: 0px 0px 4px rgba(0, 0, 0, 0.2);
-}
-
-.pagenumbers button:hover {
-	background-color: #44EEAA;
-}
-
-.pagenumbers button.active {
-	background-color: #44EEAA;
-	box-shadow: inset 0px 0px 4px rgba(0, 0, 0, 0.2);
-}
