--- PAGINACIÓN/producto_paginación.html
@@ -1,82 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-
-<head>
-	<meta charset="UTF-8" />
-	<meta name="viewport" content="width=device-width, initial-scale=1.0" />
-	<title>Users</title>
-
-	<link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/css/bootstrap.min.css" rel="stylesheet"
-		integrity="sha384-Zenh87qX5JnK2Jl0vWa8Ck2rdkQ2Bzep5IDxbcnCeuOxjzrPF/et3URy9Bv1WTRi" crossorigin="anonymous">
-	<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js"
-		integrity="sha384-OERcA2EqjJCMA+/3y+gxIOqMEjwtxJY7qPCqsdltbNJuaOe923+mo//f6V8Qbsw3"
-		crossorigin="anonymous"></script>
-
-	<link rel="stylesheet" href="../CSS/estilos.css" />
-
-	<link href="https://fonts.googleapis.com/css2?family=Bebas+Neue&display=swap" rel="stylesheet" />
-	<link rel="stylesheet" href="/PAGINACIÓN/producto_paginacion.css" />
-</head>
-
-<body style="background-color:black ;">
-
-	<!-- NAVBAR -->
-	<nav id="navbar"></nav>
-	<!-- NAVBAR -->
-
-	<!-- HERO -->
-	<section>
-		<div class="container">
-			<div class="row row-cols-2">
-				<div class="col text-center">
-					<img class="imagenhoodie1" src="../assets/img/NEW HOODIES PRINCIPAL.png" style="width: 70%"
-						alt="..." />
-				</div>
-				<div class="col d-flex flex-column align-items-center justify-content-center">
-					<h2 class="text-end-center text-uppercase font-size-lg">New Hoodies</h2>
-					<span class="text-uppercase">Flash Sale</span>
-				</div>
-			</div>
-		</div>
-	</section>
-	<!-- HERO -->
-
-	<!-- ENCABEZADO -->
-	<section>
-		<hr>
-		<div id="tituloHoodies2">
-			<h1 class="text-center fs-1 text-uppercase">Colecciones</h1>
-		</div>
-		<hr>
-	</section>
-	<!-- FIN ENCABEZADO -->
-
-	<!-- LISTA -->
-	<main>
-		
-		<div class="list" id="list"></div>
-		<div class="pagination" id="pagination"></div>
-		
-		
-	</main>
-	<!-- LISTA -->
-
-	<!-- FOOTER -->
-	<footer id="footer"></footer>
-	<!-- FIN FOOTER -->
-
-	<!-- SCRIPT PAGINATION -->
-	<script src="/PAGINACIÓN/producto_paginacion.js"></script>
-	<!--FIN SCRIPT PAGINATION  -->
-
-	<script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.2/dist/js/bootstrap.bundle.min.js"
-		integrity="sha384-OERcA2EqjJCMA+/3y+gxIOqMEjwtxJY7qPCqsdltbNJuaOe923+mo//f6V8Qbsw3"
-		crossorigin="anonymous"></script>
-	<script src="../JS/productos.js"></script>
-	<script type="module" src="../JS/navbar.js"></script>
-	<script type="module" src="../JS/footer.js"></script>
-
-
-</body>
-
-</html>
